+++ CondTools/RPC/src/RPCFebmapSH.cc
@@ -1,8 +1,8 @@
 /*
  *  See headers for a description
  *
- *  $Date: 2009/04/13 20:47:49 $
- *  $Revision: 1.1 $
+ *  $Date: 2009/06/05 15:57:03 $
+ *  $Revision: 1.3 $
  *  \author D. Pagano - Dip. Fis. Nucl. e Teo. & INFN Pavia
  */
 
@@ -18,8 +18,8 @@ popcon::RpcDataFebmap::RpcDataFebmap(const edm::ParameterSet& pset) :
   host(pset.getUntrackedParameter<std::string>("host", "source db host")),
   user(pset.getUntrackedParameter<std::string>("user", "source username")),
   passw(pset.getUntrackedParameter<std::string>("passw", "source password")),
-  m_since(pset.getUntrackedParameter<unsigned long long>("since",5)),
-  m_till(pset.getUntrackedParameter<unsigned long long>("till",0)){
+  m_since(pset.getUntrackedParameter<uint64_t>("since",5)),
+  m_till(pset.getUntrackedParameter<uint64_t>("till",0)){
 }
 
 popcon::RpcDataFebmap::~RpcDataFebmap()
